+++ .eslintrc.js
@@ -1,1 +1,3 @@
-index.js
+const { parserOptions, ...config } = require('./index');
+
+module.exports = config;