+++ CHANGELOG.md
@@ -5,22 +5,42 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 2.1.0 - 2020-08-31
+## Unreleased
+
 ### Changed
+
+- **Breaking change:** This config is now for TypeScript only
+- This config is now based on the following presets for simplified setup:
+  - `eslint-config-airbnb`
+  - `eslint-config-airbnb-typescript`
+  - `eslint-config-prettier`
+
+## 2.1.0 - 2020-08-31
+
+### Changed
+
 - **Rule change:** Disable `react/jsx-no-undef` in TypeScript files
 
 ## 2.0.0 - 2020-08-08
+
 ### Changed
+
 - **Rule change:** Stricter import ordering, including alphabetic ordering (auto-fixable)
 
 ## 1.0.2 - 2020-07-26
+
 ### Added
+
 - Documented using the Prettier configuration
 
 ## 1.0.1 - 2020-07-26
+
 ### Fixed
+
 - Fixed package Node dependencies
 
 ## 1.0.0 - 2020-07-26
+
 ### Added
+
 - **The first set of rules!** This is the first public release of this ESLint configuration.